// types/patients.ts
export type patientData = {
  name: string;
  gender: string;
  age: number;
  image: string;
};

export type patient = {
  name: string;
  gender: string;
  age: number;
  profile_picture: string;
  date_of_birth: string;
  phone_number: string;
  emergency_contact: string;
  insurance_type: string;
};